debugging.
tweaked some gesture usability features.  it should now be easier for the player to move blocks slightly and slowly to the side as the block continues to fall without the gesture type being modified mid-way through.  also, this does not resort to the previous problem of allowing the player to actually "drop" the block backward to a previous position.

added some music files.